+++ PerimeterXModule/DataContracts/Requests/RiskRequest.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace PerimeterX
+{
+	[DataContract]
+	public class RiskRequest
+	{
+		[DataMember(Name = "request")]
+		public Request Request;
+
+		[DataMember(Name = "vid", EmitDefaultValue = false)]
+		public string Vid;
+
+		[DataMember(Name = "additional", EmitDefaultValue = false)]
+		public Additional Additional;
+	}
+}
