+++ PerimeterXModule/DataContracts/Requests/RiskRequestHeader.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.Serialization;
+
+namespace PerimeterX
+{
+	[DataContract]
+	public class RiskRequestHeader
+	{
+		[DataMember(Name = "name")]
+		public string Name;
+
+		[DataMember(Name = "value")]
+		public string Value;
+	}
+}
