+++ PerimeterXModule/DataContracts/Responses/Helpers.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace PerimeterX
+{
+    public class Helpers
+    {
+        public Helpers()
+        {
+        }
+    }
+}
