+++ PerimeterXModule/PerimeterXModule.csproj
@@ -88,6 +88,15 @@
     <Compile Include="PxModuleConfigurationSection.cs" />
     <Compile Include="DataContracts.cs" />
     <Compile Include="ActivityReporter.cs" />
+    <Compile Include="DataContracts\Requests\RiskRequestHeader.cs" />
+    <Compile Include="DataContracts\Requests\Request.cs" />
+    <Compile Include="DataContracts\Requests\Additional.cs" />
+    <Compile Include="DataContracts\Requests\RiskRequest.cs" />
+    <Compile Include="DataContracts\Responses\RiskResponseScores.cs" />
+    <Compile Include="DataContracts\Responses\RiskResponse.cs" />
+    <Compile Include="DataContracts\Requests\CaptchaRequest.cs" />
+    <Compile Include="DataContracts\Responses\CaptchaResponse.cs" />
+    <Compile Include="DataContracts\Activity.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="block.mustache" />
@@ -105,6 +114,10 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup />
+  <ItemGroup>
+    <Folder Include="DataContracts\Requests\" />
+    <Folder Include="DataContracts\Responses\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\..\SimpleWebApp\packages\NuGet.Build.Packaging.0.1.248\build\NuGet.Build.Packaging.targets" Condition="Exists('..\..\SimpleWebApp\packages\NuGet.Build.Packaging.0.1.248\build\NuGet.Build.Packaging.targets')" />
 </Project>
